Refactor Hompage code - Make code connected. - Split into more parts.
