+++ js/utils/post.js
@@ -1,0 +1,50 @@
+import dayjs from 'dayjs'
+import relativeTime from 'dayjs/plugin/relativeTime'
+import { setTextContent } from './common'
+
+// to use relativeTime
+dayjs.extend(relativeTime)
+
+export function createPostItem(post) {
+  if (!post) return
+
+  // find template post
+  const postTemplate = document.getElementById('postTemplate')
+  if (!postTemplate) return
+
+  const liElement = postTemplate.content.firstElementChild.cloneNode(true)
+  if (!liElement) return
+
+  // update title, desc, author, ...
+  const thumbnailElement = liElement.querySelector('[data-id="thumbnail"]')
+  if (thumbnailElement) {
+    thumbnailElement.src = post.imageUrl
+
+    thumbnailElement.addEventListener('error', () => {
+      thumbnailElement.src = 'https://via.placeholder.com/1368x400?text=Image+not+found'
+    })
+  }
+
+  setTextContent(liElement, '[data-id="title"]', post.title)
+  setTextContent(liElement, '[data-id="description"]', post.description)
+  setTextContent(liElement, '[data-id="author"]', post.author)
+  setTextContent(liElement, '[data-id="timeSpan"]', `- ${dayjs(post.updatedAt).fromNow()}`)
+  // attach events
+
+  return liElement
+}
+
+export function renderPostList(postList) {
+  if (!Array.isArray(postList)) return
+
+  const ulElement = document.getElementById('postList')
+  if (!ulElement) return
+
+  // clear current list
+  ulElement.textContent = ''
+
+  postList.forEach((post) => {
+    const liElement = createPostItem(post)
+    ulElement.appendChild(liElement)
+  })
+}
